feat: support changing the logger and log level at runtime WPB-11541
Currently it's only allowed to set a logger and log level once at runtime,
this restriction is annoying for clients since they setup and teardown the
core crypto stack when switching accounts. It also makes it harder than it
should be to have developer setting for increasing the core crypto log level.