fix: avoid spaces in kotlin test names
Kotlin apparently lets users name functions with arbitrary strings,
as long as they are enclosed with backticks. Unfortunately, this
feature is apparently not supported in the Android version of the JVM,
causing the test suite to fail to run on Android.

A brief examination of peer repositories maintained by Wire suggests
that arbitrary-string function names are unusual. So to enable the
Kotlin test suite to potentially run on Android, this PR renames
all affected functions.