feat: relax `Debug` trait bound on `CoreCryptoCommand` and add Rust helper [WPB-12132]
There is no good syntax for implementing a trait
inline on an anonymous struct in Rust, so actually using the
`transaction` API was kind of a pain. There _is_ good syntax for
defining lambda expressions inline, so it would be useful to
add a blanket implementation allowing a lambda to be used as
a `CoreCryptoCommand`.

Unfortunately, lambda expressions and functions do not implement
`Debug` by default, making this impossible while `CoreCryptoCommand`
is a subtrait of `Debug`.

Fortunately, it turns out that we did not actually use the `Debug`
trait anywhere on `CoreCryptoCommand`, so it was kind of pointless.
So this commit relaxes that subtrait bound and adds a helper
such that lambdas can be used within transactions.