feat: add helper to extract data from within a transaction
The new transaction interface is primarily designed for FFI. On the
positive side, it's simple. On the negative side, it requires interior
mutability in order to get any data out of a transaction.

That pattern is complicated, so this commit adds a helper which can
accomplish that with safe Rust code.