build: replace `async-lock` with `tokio` locks in `core-crypto`
This is experimental work, the purpose of which is to see how hard
it would be to just swap out the lock implementations. If all had
gone well, then it would have been worthwhile to make this change
for the `map` and `try_map` that `tokio`'s lock guards have, which
could have improved the ergonomics of some of our internal methods.

Unfortunately, the difficulty was more than trivial. While it took
only a few minutes of mechanical work to swap out the imports and
uses, this causes a handful of tests to fail. So it's not really
worth proceeding further with this experiment until we've actually
prioritized it.