build: add an API with which to retrieve build metadata [WPB-11001]
This data is stringly typed due to passing through the environment
between being generated and being incorporated into the program.
No attempt is made to parse it in order to keep things simple;
the expected use case for this data is just log information anyway.

This adds more information that was strictly requested by the PR,
in the assumption that this information is negligibly cheap to store and
can only help with reducing parsing demand.