feat: proteus error codes are `Option<u16>` not `u32`
Returning an error code whose zero value indicates success is
a reasonable pattern, in C.

We are not programming in C.

Eventually we're going to get rid of this pattern entirely. For now,
we can at least use a real `Option` type to indicate whether an error
in fact exists.

Also, the Proteus implementation specifies that the error code is
a `u16`. No idea why that got widened to a `u32`, but that was
pointless, so it's back to a `u16` here.