chore(uniffi/jvm): get kotlin to build _at all_
Kotlin tests are still failing due to previous changes, but at least
the fundamental build operation is now succeeding. It's a start!

Ran a Kotlin code formatter, intending only to hit the auto-generated files;
turns out there were a bunch of human-maintained files in the tree also.
Unfortunately, at that point, there were quite a lot of uncommitted changes.
Didn't want to go through the effort of going back and unformatting the
unrelated stuff, so that's all in here also.