refactor!(e2e_identity): module uses a module-specific error type
By always returning a local error type, we open the door to removing
the centralized error types entirely in favor of a more specific error stack.

Note that this commit introduces a deprecated wrapper aroud `CryptoError`
within `e2e_identity::error::Error`. This will be removed shortly, but
it is necessary as an interim step, so that we don't have to rewrite
actually everything, all at once.

Note also that this affects some files within `mls`. This is required
in order to keep everything buildable.