refactor(crypto-ffi): `core-crypto-ffi` builds given new errors
The error-handling status quo in `core_crypto` is wildly different now,
so we needed to change quite a lot about how `core_crypto_ffi` used
those errors.

This commit causes `core-crypto-ffi` to build properly.

Weirdly, this was _much_ simpler in wasm than doing so for non-wasm. I'll take
it though!