chore!: don't restore pending messages when own commit was accepted
Now there is no visible to the outside state in core crypto representing the gap between our own commit being created and the commit being accepted. There will no messages be received in between. This is why we don't always need to restore pending messages anymore when accepting a commit. There are still some cases where messages will need to be restored, which will be handled in `decrypt_message()`.