fix: fix decryption of pending messages when receiving own commit
Previously, pending messages weren't completely processed when decrypting an own commit message.
This wasn't noticed because previously the intended usage was to call `commit_accepted()` when a pending commit could be merged. `commit_accepted()` did the decryption of pending messages.
`decrypt_message()`, left pending messages untouched if the message was an own commit.

This needed fixing now, because now we rely on `decrypt_message()` to process pending commits.