chore: decrypt message: merge pending group from external commit
This was previously handled by explicit calls to `merge_pending_group_from_external_commit()`. However, this method is not exposed anymore, so we need to handle it inside decrypt_message(), for the case that we didn't receive the success response during the send_commit() call (e.g., because of application crash).