fixup: `innermost_source_matches` can handle dereferencing boxed errors
Basic error reporting and debug representations of things hide `Box`es.
It's also not trivial to downcast into them, particularly in the context
of a macro like this!

Non-trivial does not mean impossible, though. Adding this capability
to the macro means that we can avoid an ugly workaround in the test suite.