test: remove tests about "leaking entities"
Some methods where annotated with an _idempotent_ or _dispotent_ attribute.
This would add to all calls in tests an assertion that calling them did never or did always change the count of records in the database.

This is no longer the case when the mls transport API is used.
Now, depending on the result of an MLS transport api call, a function call may or may not insert or delete a record, which makes this attribute obsolete.