build: adjust wasm builds
After a fair amount of time wrestling with the WASM builds, I've
come to the conclusion that centralization is just not worth it.
It does in principle save several minutes of build time by doing
the hard part once instead of four times, but after several days
of effort, it's just not working. The tools we need to use are
kind of the opposite of idempotent.

More details here: https://github.com/wireapp/core-crypto/pull/823#issuecomment-2594790198