fix(web): broken error type mapping in try/catch patterns.
Apparently the `CoreCryptoError.fromStdError()` function does not work as expected.

We can easily fix error type mapping by doing an early return at least for the cases
where this function is called with an error that is already an instance of `CoreCryptoError`.

Now calling this function at least doesn't make things worse.