refactor: don't return `Err(Error::BufferedFutureMessage)` from `handle_future_message()`
This was meaningless because in the one and only caller this error would always be propagated.
So we can just keep it in that context and raise it from there directly.
No need for an extra round trip.

(cherry picked and adjusted from commit 5d7af36053ca4921ee40b759f0423206e73215c1)