chore: create and link Jira issues for `TransactionWrapper`
Turns out that we are doing some pretty naive/unsafe things with
regard to `TransactionWrapper`. Ultimately this comes down to the
fact that we're using the wrong tool for the job: the advice
from the maintainers of rusqlite regarding async is "don't use this
tool in that context."

In the meantime, we can't easily change any of these problems...
but we can at least document them!