refactor: ensure safe use of `unsafe` for `SqlCipherConnection`
There used to exist a safety note for the unsafe impls of auto-traits
for `SqlCipherConnection`. Unfortunately, it straight-up lied about
the properties of the struct.

Fixing this involved adding a mutex, which required removing the
`Deref` and `DerefMut` implementations, which then propagated into
some moderately widespread code changes. Nothing too interesting.