chore: crypto-ffi: don't use the base maven-publish plugin
The base plugin is missing tasks to publish to Maven Central.
Instead, use the high-level plugin, that provides those tasks,
but still try to disable javadoc generation.

For POM information it looks like the high-level plugin picks
up gradle.properties automatically, so we can remove the

  pomFromGradleProperties()

bit. Actually, we have to remove it, because it's incompatible with the
high-level plugin and using it results in an error.