test(buffer_messages): add test case for the first part of 15810
This test case replicates the 15810 scenario:
the observer and new member process different proposals removing
member 114, but they can still talk to each other.

Member 27 receives these messages in another order entirely,
and generates its own commit based on that. However, it is
possible for other group members to fail at this depending
on the order that they receive the proposal and the commit.