chore: interop: don't import UniffiCustomTypeConverter
uniffi has removed UniffiCustomTypeConverter, and it's not
possible to use uniffi::Lift::try_lift because we would need
to instantiate a proper uniffi RustBuffer for our type,
and that in turn requires us to have an instance of our
type, which we don't -- we're trying to construct such
an instance.

So just use the From impl to construct the wrapper ClientId instance.