chore: ci: use stable chrome version
This is a revert of 54f9a8642a7ca2196800e19517b052bc2eca1ea6, which we
had to do for a good reason. However, now we run into
https://github.com/browser-actions/setup-chrome/issues/604.

So let's try stable again and hope it works.
This also makes it consistent with our other jobs that use chrome stable.