test: don't assert a non-epoch-change when the epoch must change
The assertion removed here replaced an older assertion against the
now-deprecated attribute `has_epoch_changed`. However, that assertion
made no sense in this new context: the point of this test case is to
ensure that `decrypt_message` is durable, and there are other
assertions here that the number of members in the conversation has
changed. So of course the epoch must change!