refactor!: move mls transport structs into a unified module
Unfortunately, uniffi and wasm_bindgen handle external interfaces
in completely different ways which are structurally incompatible
with each other. Ultimately we end up defining things separately
instead of even attempting compatibility.

Worse, in wasm-land, you can't fully specify the proper type signature
for an imported async function. The best you can do is declare the
unchecked typescript return type, but even that requires us to upgrade
our wasm_bindgen to a version which our CI has had problems with in
the past.

We'll get to that problem later. For now, this commit represents the
best I could do after several hours of wrestling with uncooperative
wasm bindgen errors.

Incidentally, this completely changes the interface used to register
the Mls transport API on the JS/TS side of things. This should be
both simpler and easier to understand, but it will definitely break things.