refactor!: move CoreCrypto definition, constructors to a common module
This is setting up the necessary changes to organize the whole CoreCrypto
method implementations into this subtree. There are a whole bunch of
trivial changes because even at this point, the CoreCrypto definitions
were not precisely the same between uniffi and wasm.

This commit unfortunately worsens the error message that we get when
trying to close out the CC instance when there were multiple strong
refs remaining. For now I'm going to assume that that's an unusual
enough case that we don't mind too much. In exchange, we have 1 fewer
pointer to chase every time we do anything; that might be detectable
if we set up the right microbenchmark!