chore: keystore: drop support for legacy keychain names for db salt
Originally we supported just one salt, under the name "keystore_salt"
and then we allowed storing multiple salts by making the name include
the hash of the database path, resulting in "keystore_salt_{hash}".

All of this is pre-1.0 CoreCrypto, so we can safely drop support for the
original single salt entry.