fix(wasm): WIP duck types are not trait objects
You just have to bite the bullet and store and manually call
`Function` instances.

There's more coming for this commit, as there are a few places where we used
the duck type pattern that we now have to excise.

- [x] logger
- [ ] epoch observer

Note that it's probably fine to retain duck typing for `CoreCryptoCommand`
as we don't store it, but just execute it all the way through.