fix(wasm): duck types are not trait objects
You just have to bite the bullet and store and manually call
`Function` instances.

This updates the following instances where previously we were attempting
to use duck-typed instances as trait objects:

- [x] logger
- [x] mls transport
- [x] epoch observer

Note that `CoreCryptoCommand` is not part of that set: in that case,
we don't store the instance beyond the duration of a single function
call. That's one of the explicitly-supported use cases for duck types!
So we leave it, there; where we can use it, the usage patterns are
just nicer than calling Function objects.