refactor!: move logger structs into a common module
Once again we find an interface where what we _needed_ was a duck-type interface,
and what we _had_ was some kind of manual implementation doing things by hand,
and it did not line up at all.

So this doesn't just move all the stuff; it also once again does the duck-type
dance that wasm needs.