refactor: rename `mls::client::Client` to `mls::session::Session`
`Client` is a bad name because it's overloaded: in one sense it's
what the MLS RFC uses to mean our software interacting with the
mls protocol. In another sense it's the consumer of the library
that we publish. In a broader sense it's the union of Wire software
on a user's device: both our library and the user-facing bits that
call into our code. It's very challenging to speak precisely
using this term.