fix: ci: fix permission issues with the container
There are two fixes here.

Firstly, the container user (`step`) must have write access to the directory that is
bind-mounted from the host in order to be able to write certificates to
disk. So we just set mode 0777 on the bind mount dir.

Secondly, the host user (the one outside the container, running the test) needs permission
to read certificates generated by ACME, as well as a permission to alter ACME's configuration
(this happens outside the container).