fix(test): `join_by_external_commit` fails when given an invalid group info
Once upon a time, the JBEC method validated the age of a keypackage,
so this test constructed a groupinfo containing an expired keypackage.

Then later that got disabled for Reasons, and this test was updated to
do nothing except waste 15 seconds. Some time after that, the relevant
issue was created to make the test do something again.

Ultimately, we just needed an invalid `VerifiableGroupInfo`.
It did not seem worth creating a helper method for constructing
an invalid one, but we got around it by inlining the call which
constructed the groupinfo and adjusting its parameter appropriately.