refactor(test conversation): use a type-state-like pattern
This solves the problem that we couldn't have a mutable and immutable
reference to the test conversation at the same time in a test.
Now, in the places were we had a mutable reference before, we consume
the conversation and return it after modification.
As a side-effect, the usage patterns are nicer, too.

Also, rename `invite_and_notify()` to `invite()` and introduce
`invite_guarded()` to replace the old `invite()`. Now the more
frequently used function has a shorter name.