refactor!(ffi): make `NewCrlDistributionPoints` a typedef not a newtype
Our high-level downstream interfaces undo the newtype conversion anyway,
presenting it as a typedef. So there's no point in retaining this newtype.

Eliminating this also means that `DecryptedMessage` and `BufferedDecryptedMessage`
don't need redefinition in the high-level bindings anymore; they are structurally
equivalent to the generated version. This opens the door to removing those
in the future.