refactor!(js): use `ExternalSenderKey` newtype
Affected items:

- `getExternalSender` returns this newtype
- `ConversationConfiguration` accepts this newtype

In addition to the obvious benefits of having a strong type instead
of an arbitrary byte buffer, this has the additional benefit that
in the expected case, the keys will never actually be transferred
across the FFI boundary. Instead, they live on the Rust side of things
at all times, making affected operations somewhat cheaper.

This opens the door to removing the high-level definition entirely, as its
interface now exactly matches that of the Rust/FFI version.