refactor!(kotlin): use `ConversationId` inside `MLSGroupId`
This sets us up for a later step in which we get rid of `MLSGroupId`
entirely in favor of using the exposed FFI type.

This would be an invisible change to the user except that in the
`HistoryObserver` and `EpochObserver`, we were previously exposing
byte arrays instead of any kind of wrapper. This fixes that and
exposes the existing `MLSGroupId` wrapper in those interfaces instead.