fix: proteus error type mapping
Until now, we incorrectly assumed, that we could determine whether a
recursive error is a proteus error by just looking at the innermost
type. However, this approach fails in the case of a
`ConversationNotFoundError`, wrapped in a `LeafError`, which is not
proteus-specific.

Therefore, we're now checking entire error stack for proteus error
kinds.