feat(swift): hold a file lock on they keystore while executing a transaction
Protect against multiple core crypto instances accessing the same key store by
holding a file lock on it during a transaction.

The iOS app already does but it's much nicer if we do this internally since it
removes the need for the iOS app to build an abstraction over core crypto.