refactor(ffi)!: uniffi exports native Time and Duration instances
... instead of integers which need further interpretation.

wasm-bindgen is not capable of exporting these abstractions, so there
is some necessary API variation between the inteface exposed in wasm
vs in swift or kotlin.

This change further eliminates the need for high-level wrappers,
which previously existed in Kotlin for the purpose of converting between
integers and these higher-level time and duration instances.

This is a breaking change in swift!