refactor(jvm)!: eliminate pointless redefinition of equivalent FFI types
This means that to a much larger extent than previously, Kotlin callers
will be working directly with FFI types.

This is not perfectly semantically equivalent; there are instances where
for example the order of fields in a function parameter list have changed.
Also, no functions now have any default parameters. However, in general,
types and functions should be familiar even if they are not identical.