docs: remove `7.x` series from `index.md`
Docs weren't built and deployed to the correct directory. We decided
that manually building and uploading docs for every platform, which
would be necessary to fix this, isn't worth it. Not providing docs for
`7.x` will be a good nudge for application developers to migrate to
`8.x` as soon as possible.

Also, fix a typo in the rust column.

(cherry picked from commit 6102d9fa5c364c18d2a8439cb60d381e11f4b298)