fix: warning about DYLIB_CURRENT_VERSION getting truncated
When build the Swift bindings we get the following warning:

```
warning build: Truncating -current_version to fit in 32-bit space used by old mach-o format
```

This is because the DYLIB_CURRENT_VERSION is now larger (80000) than 65535, which is
the calculated as major * 10000 + minor * 100 + patch.

We shouldn't convert our version into single integer but rather keep it as is since
the acceptable format for this version is xxxx.yy.zz.