chore(test): add implicit test executor attributes
In rstest v0.26, the framework no longer guesses that you have
async_std as a peer dependency. Instead, you must specify the
appropriate async test attribute either implicitly or explicitly.

Here, we add the implicit test attributes. These are precisely the
places we'll need to update to explicit test attributes when we
move to smol.