test(keystore): fix previously-invisible test failures
There were two problems with the keystore wasm tests:

1. The `Drop` implementation for `KeystoreTestContext` used `block_on`,
   which is not allowed in wasm. You can't just block on things there;
   it's disallowed.
2. There was one test which could not pass, as it checked for an error
   which was only ever produced by wrapping a `rusqlite` error.

However, both of these problems were being swallowed somehow when
being run on wasm, and `wasm-pack` was happily reporting every test
unconditionally as passing regardless of what was happening on the
inside. Given that this problem was revealed when we swapped out
`async-std` for `smol`, I'm going to go ahead and blame something
in `async-std` for this. I'd open an issue, but `async-std` is now
unmaintained and deprecated, which is how we even found out about this.

Checklist for supporting evidence:

1. Check out the branch `prgn/experiment/DO-NOT-MERGE-wasm-test`
2. Run `wasm-pack test --headless --chrome ./keystore --test z_entities -- --nocapture update_e2ei_enrollment_emits_error`
3. Note that the output includes both the `starting test` and `getting error`
   lines, but _not_ the `ending test` line.