build: get rid of `release-build` target
We don't need a step letting us explicitly build `core-crypto-ffi`;
we'll just invoke `cargo` directly if we want that. Depending on
the `release-build` step was messing with our build process, as well:
it was marked as phony, meaning that everything downstream was marked
as dirty even if cargo had decided that it was all fine.

Getting rid of it means that make knows what file to actually look at,
and can rebuild or not based on that.

Also we have the `$(UNIFFI_BINDGEN)` variable, so use it.