build: move docs targets later in the file
Apparently `make` is sensitive to the order in which things are defined,
and `make docs-ts` wasn't doing the right thing previously becuase
its dependency `$(DTS_OUT)` wasn't yet defined at check time,
even though it was defined later when it was being executed.

This puts the docs tasks after all relevant variables are defined,
so now it behaves as expected.