ci: interop: use one simulator device for interop
We used to create a device for building the App used in the test and
another device to actually run the test. We can do both with the same
simulator device.
Note that a device deletion will shutdown the device if it is booted.
Thus, we remove the manual shutdown of the device.