ci: continue if an uploaded artifact already exists on a run
This is a workaround for actions/upload-artifact#506 (overwrite behavior
is not atomic). The reason why we're getting in this situation is that
we want to use the same composite action whenever we want to guarantee
that an artifact exists. This makes our workflows more robust in cases
where the artifacts that were uploadded by different workflow runs have
expired (e.g., a retry several days later).