ci: remove targets list from rust-toolchain.toml before rust setup
The setup-rust-toolchain action will install all targets specified in
the rust-toolchain.toml even if we specify a target as an argument. This
wastes CI runtime. Therefore we remove the targets list before running
the action. Note that on macOS gnu-sed is called via gsed.

Explanation of the regex:
^targets = \[ → match the beginning of the targets array.

,^] → up to and including the closing ].

d → delete that whole range.

-i → edit in place