feat: define obfuscate trait and Obfuscated type
Main purpose for this trait is to define the obfuscate function for any
type that should be obfuscated when debugged.
To make use of the formatter when using a custom derive debug macro, we
define the wrapper type Obfuscated which implements fmt via obfuscate.

We reuse the implementation from the old obfusctaed trait and already
take definitions for external types from openmls.