feat: add #[sensitive] attribute to debug macro
When a field or struct has the attribute #[sensitive] then its content must not be
logged as plain text. Therefore we use the Obfuscated wrapper which
calls obfuscate on fmt. Each field that is sensitive just has to
implement the Obfuscate trait.
All fields of a sensitive struct are considered sensitive.
To enable structured key value logging by default we also impl log::kv::ToValue via the macro.