refactor: obfuscate for HistorySecret
Note that HistorySecret::obfuscate can rely on the Debug implementation
of ClientId which marks the id as sensitive. KeyPackageEncapsulation on
the other Hand is an external type. Here we have to use our wrapper or
use a newtype if we want to use the debug macro with sensitive. We opt
for the wrapper as done before.