build: adjust rustdoc-wrapper.sh to better strip GFM
Previously it stripped GFM admonitions written in lowercase.
However, strictly, they're supposed to be in uppercase, and there
is a somewhat larger set than were previously supported:

<https://github.com/orgs/community/discussions/16925>

As we now have a markdown format checker in CI, we know that these
admonitions will be rendered in uppercase, so we had to adapt the
sed expression. We updated the set of admonitions en passant.