build: Makefile: add proper dependencies for docs-kotlin
We can't depend on `jvm` target because that's phony and will cause docs-kotlin
to trigger always -- we must depend on its output, which is $(JVM_LIB).

We also need to add docs-kotlin-deps because of CI, which expects -deps
rules in order to be able to calculate hashes.

Finally, we also need to consider Kotlin sources when listing dependencies.