build: use nightly toolchain for dev
This change is to enable nightly features. There are some very useful
nightly features which don't make into stable for years, e.g. import
stucture for rustfmt.

We decided to give this hack a try: We pin the rust toolchain to nightly
and set the rust-version of each crate to the used stable version. This
ensures compatibility with any stable build process but allows us to use
nightly features during development. Consequently, we'll use nightly
linting and formatting in CI.