refactor: `MlsCredentialType` -> `CredentialType`
It's crazy how bad the naming scheme had been. Sometimes in the past
we had `MlsFoo` meaning "the foo entity in the keystore"; in this
case, we had `MlsCredentialType` meaning "the credential type entity
defined in core-crypto."

So we change it: places where we need to reference the Openmls
credential type use it as `MlsCredentialType`. Places using our
own credential type just call it `CredentialType`.