refactor: drop `MlsCredentialType`
It provided us exactly one thing: it didn't handle the "unknown" case,
because it incorrectly featured a bug which would trigger a runtime
panic if one was ever encountered.

That was obviously not ideal, so it seemed best to simplify the
type system by dropping it entirely in favor of the upstream type.