chore: WIP rm old credential API
Unfortunately I just ran out of time here. Certainly parts of the API
are now gone, but I could not complete a full audit of all the old code
to ensure I got rid of all the old credential entry points.

In addition, I had to resort to some `todo!()` items in the test suite;
it certainly does not pass at the moment.

Key unresolved questions:

- How can we recover a `ClientIdentifier` from a session / test case in x509?
- Should we even have a `ClientId` as part of `ValidatedSessionConfig`?
- If not, can we improve the initialization API?
- Are there any more bits of newly obsolete credential management code floating
  around waiting to be removed?