feat!: rm `SessionConfig` and `ValidatedSessionConfig`
These were type-safe wrappers around 3 fields: a database,
a client id, and a list of ciphersuites. Of these, the latter
two were responsible for all the validation requirements, and
were only used to instantiate credentials with the non-obvious
code path "if a client id is provided, mls gets initialized."

Part of the point of this whole PR is that credential creation is
always explicit, and session instantiation is less magical. So
we just ripped out both config instances. Sessions now require
a Database instance to instantiate, and no other parameters.
They never automatically instantiate MLS.