feat: order in which credentials are added no longer matters
It used to be the case that credentials could only be added to the
identities list in strict order by timestamp. This was because
`Identities::find_most_recent_credential` was based on a data
structure which stored the most recently _added_, not the actual
most recent credential.

Obviously this isn't a viable restriction to keep if we're going to
decouple credential management from initialization, so we needed a
better data structure. I chose to implement this with a strong ordering
on credentials, plus a BTreeSet which ensures we can get the most
recent credential on demand without significant overhead.

This comes with a new restriction: because we're no longer tracking
insertion order but the actual fields of the credential, it is no
longer possible to add two credentials which have the same signature scheme,
credential type, and earliest validity. This is because our identities
set no longer can usefully distinguish between those credentials, and will
return as the most recent whichever of them has the greater identity
when expressed as a byte slice. Appropriate runtime checks are added
to prevent this.