feat: `CredentialRef` keeps track of earliest validity
This gives us more accurate db queries when loading the credential
ref and exposes a little bit more cheap information to the clients.

Ultimately the reason why we have a restriction on adding multiple
credentials in the same second is that the database can't distinguish
between them. The database gives us 64 bits of precision, as seconds.
If that weren't the case, or migrations were easy, we'd just up the
resolution to nanoseconds and trust that computers aren't _that_ fast.

But because the database has a timestamp resolution of 1s, we can't
store any greater than that in the credential itself, because on load
the database gives us our credentials back in arbitrary order and we
must construct the appropriate ordering for them ourselves.

Keeping track of which second that _was_, in a way visible to clients,
can only improve things.