chore: explicitly save the created credential in `e2ei_mls_init_only`
E2ei is hideously complicated and at least some of that complication
is because credentials used to be _very_ finicky. I'm intentionally
trying not to mess with any of this more than I have to.

This is a case where I had to. IDK how the credential used to be
persisted, but these days, once there is a credential it is easy enough
to explicitly persist it into the keystore.