fix: `CredentialRef` includes the credential's public key
Without this, it is very difficult to line up the retrieved
signature key pair with the retrieved credential. It's just
another thing we'd want to fix in the future when we have
a more sane way of interacting with the database.

With that sorted, I could remove/refactor a few more obsolete
credential APIs.