feat: add `Drop` impl to wasm storage; rm `.close()` methods
It looks like ultimately the only concrete benefit that `.close()`
gave us was to ensure that on WASM, storage got flushed at the end.
Rusqlite already had that behavior.

It turns out to dramatically simplify the API if we don't need to
be able to drill down and take ownership of the most fine-grained databse
connection from way out at the FFI level, so this commit just
adds a `Drop` impl to the wasm storage wrapper and removes all the
`.close()` impls.