feat: enrollment never needs exclusive access to unwrap an arc
Right now we have a category of error caused by the assumption that
we have exclusive access to an `Arc<Enrollment>`. This is a problem
over FFI, because foreign languages often don't have the ability
to clearly state when an Arc reference is created, or notice
when one is deleted.

So we remove that assumption/requirement from E2eiEnrollment in this commit.