build: do not run wasm tests except for the keystore
1. `wasm-pack` is unmaintained[1] and does not yet have a suitable replacement[2]
2. wasm-pack is now producing test failures without reporting what specifically has failed[3]

Those test failures are kind of interesting, because they manage to be tantalizing, and provide a little information ("these 3 of over 100 tests _did_ pass..."), but omit any actionable information about what specifically has failed, where, or why. They do this very slowly, at over 20 minutes per test run.

That said, given that all the tests pass when run in a native context, and all the higher-level wasm tests still pass, it seems very likely that this is a spurious failure, not indicative of a real problem. And spurious failures which appear interesting but can't be acted on are a recipe for wasted time. Better not to generate them at all.

We retain the wasm-pack tests for the keystore, where apparently they are still important for idb testing.
Once we have a unified database backend, we can remove those also.

[1]: https://blog.rust-lang.org/inside-rust/2025/07/21/sunsetting-the-rustwasm-github-org/
[2]: https://github.com/wasm-bindgen/wasm-bindgen/issues/4634
[3]: https://github.com/wireapp/core-crypto/actions/runs/19041040144/job/54378042141?pr=1522