fix: in js 0 is falsy, which messes with ternary logic
We have this helper function which converts `bigint` to `number`
or raises an error in the unlikely case that it's out of bounds.
And for typing reasons, we want to keep its signature purely
`function(bigint): number`.

But in a few places, we needed to call that function on a value
which might be undefined. So we used ternary logic, like this:

```js
foo? saveBigIntToNumber(foo): undefined
```

And on its face that does the right thing: both `null` and
`undefined` are falsy in JS, so you get `undefined` out.

What we hadn't considered was that the defined number 0 is _also_
falsy in JS, meaning that those instances could change a defined
value to an undefined one, causing some subtle bugs.

So let's get really explicit about this and do things properly.