chore: add `KeyPackage::make_ref`
Really this is about a general strategy for improving the OpenMLS
keypackage API. Currently we're just using it all as-is. Where I
want to get to is simplifying/improving the API through the use
of an extension trait; this is quite likely to be only the first
extension method we add. But we had to start somewhere, so that's here.