feat: use namespace function credentialBasic() to create a credential
Static methods are not supported by uniffi so it generates this
namespace function. Unfortunately, we can't extend the uniffi Credential
class because its constructor is private. Therefore, we change the
client api to use the namespace function instead.

This should also be changed in swift and kotlin to align the APIs.