refactor: keystore trait: stop using `StoredSignatureKeypair`
This means that we take the keypair from the credential when reading.
Turns out that the storing and deleting a signatuture keypair is not
needed by openmls, when can just leave them `unimplemented!()`.

When reading, serializing and deserializing the openmls keypair once
is the only way I found to satisfy the compiler.