chore: allow migrating key type to bytes until v15 only
This eliminates the additional maintenance burden we'd have if we needed
to support an always up-to-date keystore dump to test this function.

This is not a breaking change, because even prior to this version it was
not possible to regularly open the database with a string-typed
key, users first had to call the migration function.