refactor: `Keypackage` is a struct of its own
We use keypackages in two ways, mostly governed by how OpenMLS uses
them: as structs emitted by our libary, and as inbound types which need
verification (which happens automatically within OpenMLS). But conceptually
they're one thing.

Meanwhile, as far as our FFI layer was concerned, keypackages were just another
typed array of bytes. We had a constructor and a `copy_bytes` method, and that
was it.

However, we wanted the ability for foreign clients to be able to do
`keypackage.ref()` and get a `KeyPackageRef`, and that was just not viable
under the previous approach. So we needed a new way of handling things.

First, we create a new struct wrapping a hidden enum. (This can likely be simplified
once we trade out OpenMLS for mls-rs.) This is the foreign type wrapping the internal
keypackage concept. At the same time, we can improve on the awkward casing.

Then we add its API. We still have a synchronous constructor, but we rename
`copy_bytes` to `serialize`. Constructor and serialize methods are both fallible,
which means that clients can recover from errors at a far more granular level now.

Usage gets simpler in `mls.rs` because it encapsulates the knowledge about how
it converts to and from a byte stream.

And of course finally, we can make public a `.ref()` method that any client can use
at any time to generate a `KeyPackageRef`.