chore: fix a test which was failing due to changed semantics
Somehow it used to be that going through an e2ei enrollment would
just automatically clear out all keypackages created with a basic
credential, despite comments in the code mentioning that this could
potentially lead to orphan keypackages.

We've removed the method which did that, so now those keypackages
are no longer removed, which meant we needed to adjust this test.