chore: REVERT / ERASE FROM HISTORY LATER
Disable the `derive(Entity)` macro to a noop.
I'm going to be doing some major refactoring of the `Entity` trait,
and the noise of compiler errors due to implementations not matching
the new definitions would absolutely drown out the signal
of real problems, except for this.

Obviously, we need to retain a real entity derive macro before this PR
is over, but this should simplify things temporarily.