fix: android-x86 target not compiling
The android-x86 target is special in that it statically links to clang_rt.builtins. The path
to these libraries were wrong, pointing to a non-exising directory in the android NDK. We fix
it by starting from the  `resource-dir` instead and navigate from there instead of `runtime-dir`
which doesn't always point the right place on all platforms.