feat: define pki environment types
On `PkiEnvironment::new()` we check if the inner PkiEnvironement can be
restored from the given database. If not the CA has not been registered,
yet.

The PkiEnvironmentProvider will be shared with the MlsCryptoProvider of
the mls session. It keeps most of its behavior for now but we will
access it mostly through the new pki environment type. This should
enable further refactoring in the future.