refactor(keystore): rework item-level encryption traits
The existing `trait EntityEncryptionExt` was confusing: it depended
on mutating the internal state, and never calling the appropriate
methods too many or too few times. It wasn't obvious when or whether
it was ever appropriate for an implementing item to override one of
the default trait methods.

In fairness, it's a complicated problem.

This commit contains my approach, which breaks it up substantially.
Entities no longer implement `Serialize` or `Deserialize`, and in fact
must not implement either of those. (If it were possible, I'd have
added negative trait bounds to that effect.) Doing this with separate
types gives us some static type safety, which is always nice!

Instead, entities which want to serialize must implement `Encrypting`,
which requires them to copy all non-sensitive data and encrypt all sensitive data
into an associated type. This associated type implements `Serialize`.

Likewise, entities which want to deserialize must implement `Decrypting`,
which requires them to copy all non-sensitive data and decrypt all sensitive data
from an associated type. This associated type implements `Deserialize`.

The benefit is that required methods have no defaults, and it is
impossible to improperly override default methods. There are now
helper traits `EncryptData` and `DecryptData`, which are automatically
implemented for all entities, and handle all the details of the
symmetrical encryption and decryption.